refactor: separate common types from role-specific types and fix CHUNK_SIZE_BYTES
