+++ common/constants.py
@@ -1,4 +1,4 @@
 """Project-wide constants (e.g., CHUNK_SIZE, default ports)."""
 
-CHUNK_SIZE_BYTES: int = 64 * 1024 * 1024  # 64 MiB default chunk size
+CHUNK_SIZE_BYTES: int = 4 * 1024 * 1024
 # TODO: Define network ports, timeouts, protocol versions.
