+++ common/types.py
@@ -1,8 +1,6 @@
-"""Shared data type definitions (FileMeta, ChunkDescriptor, ChunkRecord, etc.)."""
+"""Shared data type definitions (ChunkDescriptor, etc.)."""
 
 from dataclasses import dataclass
-from datetime import datetime
-from typing import List
 
 
 @dataclass(frozen=True)
@@ -14,16 +12,3 @@ class ChunkDescriptor:
     chunk_index: int
     size: int
     checksum: str
-
-
-@dataclass(frozen=True)
-class FileMetadata:
-    """
-    Complete metadata for a file in the system.
-    """
-    file_id: str
-    name: str
-    size: int
-    tags: List[str]
-    owner_id: str
-    created_at: datetime
